🏗️ moved constants to utils folder
