+++ src/dtos/flexible-filter.dto.ts
@@ -6,7 +6,7 @@ import {
   IsNotEmpty,
 } from 'class-validator';
 import { FlexibilityType } from 'src/enums/flexibility-type.enum';
-import { months } from 'src/constants';
+import { months } from 'src/utils/constants';
 
 export class FlexibleFilterDto {
   @IsNotEmpty()
